Re-organize package tree
