+++ src/client/java/dev/fixyl/componentviewer/config/enums/ClipboardCopy.java
@@ -22,22 +22,22 @@
  * SOFTWARE.
  */
 
-package dev.fixyl.componentviewer.option;
+package dev.fixyl.componentviewer.config.enums;
 
 import com.google.gson.annotations.SerializedName;
 
 import net.minecraft.util.TranslatableOption;
 
-public enum ClipboardFormatting implements TranslatableOption {
-    @SerializedName("sync") SYNC(0, "componentviewer.config.clipboard.formatting.sync"),
-    @SerializedName("snbt") SNBT(1, "componentviewer.config.clipboard.formatting.snbt"),
-    @SerializedName("json") JSON(2, "componentviewer.config.clipboard.formatting.json"),
-    @SerializedName("object") OBJECT(3, "componentviewer.config.clipboard.formatting.object");
+public enum ClipboardCopy implements TranslatableOption {
+    @SerializedName("component_value") COMPONENT_VALUE(0, "componentviewer.config.clipboard.copy.component_value"),
+    @SerializedName("item_stack") ITEM_STACK(1, "componentviewer.config.clipboard.copy.item_stack"),
+    @SerializedName("give_command") GIVE_COMMAND(2, "componentviewer.config.clipboard.copy.give_command"),
+    @SerializedName("disabled") DISABLED(3, "componentviewer.config.clipboard.copy.disabled");
 
     private final int id;
     private final String translationKey;
 
-    private ClipboardFormatting(int id, String translationKey) {
+    private ClipboardCopy(int id, String translationKey) {
         this.id = id;
         this.translationKey = translationKey;
     }
